requirements.txt
streamlit==1.22.0
google-api-python-client==2.86.0
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0
google-auth==2.17.3
psycopg2-binary==2.9.10
python-dotenv==1.0.0
